+++ src/core/tnlParameterContainer.h
@@ -125,7 +125,7 @@ template< class T > bool tnlParameterContainer :: AddParameter( const char* name
 template< class T > bool tnlParameterContainer :: SetParameter( const char* name,
                                                                 const T& value )
 {
-   long int i;
+   int i;
    for( i = 0; i < parameters. Size(); i ++ )
    {
       if( parameters[ i ] -> name == name )
