+++ src/diff/drawGrid3D.h
@@ -27,9 +27,9 @@
 template< class T > bool Draw( const tnlGrid3D< T >& u,
                                ostream& str,
                                const char* format,
-                               const long int i_step = 1,
-                               const long int j_step = 1,
-                               const long int k_step = 1 )
+                               const int i_step = 1,
+                               const int j_step = 1,
+                               const int k_step = 1 )
 {
    if( ! format )
    {
@@ -38,10 +38,10 @@ template< class T > bool Draw( const tnlGrid3D< T >& u,
    }
 
 
-   long int i, j, k;
-   const long int x_size = u. GetXSize();
-   const long int y_size = u. GetYSize();
-   const long int z_size = u. GetZSize();
+   int i, j, k;
+   const int x_size = u. GetXSize();
+   const int y_size = u. GetYSize();
+   const int z_size = u. GetZSize();
    const double& ax = u. GetAx();
    const double& ay = u. GetAy();
    const double& az = u. GetAy();
@@ -73,7 +73,7 @@ template< class T > bool Draw( const tnlGrid3D< T >& u,
       str << "<PointData Scalars=\"order_parameter\">" << endl;
       str << "<DataArray Name=\"order_parameter\" type=\"Float32\" format=\"ascii\">" << endl;
       str. flags( ios_base::scientific );
-      long int i, j, k;
+      int i, j, k;
       for( k = 0; k <= z_size - k_step; k += k_step )
          for( j = 0; j <= y_size - j_step; j += j_step )
             for( i = 0; i <= x_size - i_step; i += i_step )
@@ -96,7 +96,7 @@ template< class T > bool Draw( const tnlGrid3D< T >& u,
       str. put( ( char ) ( y_size & 0xff ) );
       str. put( ( char ) ( z_size >> 8 ) );
       str. put( ( char ) ( z_size & 0xff ) );
-      long int i, j, k;
+      int i, j, k;
       double min( DBL_MAX ), max( -DBL_MAX );
       for( k = 0; k < z_size; k ++ ) 
          for( j = 0; j < y_size; j ++ ) 
@@ -122,9 +122,9 @@ template< class T > bool Draw( const tnlGrid3D< T >& u,
 template< class T > bool Draw( const tnlGrid3D< T >& u,
                                const char* file_name,
                                const char* format,
-                               const long int i_step = 1,
-                               const long int j_step = 1,
-                               const long int k_step = 1 )
+                               const int i_step = 1,
+                               const int j_step = 1,
+                               const int k_step = 1 )
 {
    fstream file;
    if( strncmp( format, "bin",3 ) == 0 )
