+++ src/matrix/tnlBICGSolver.h
@@ -40,7 +40,7 @@ template< typename T > class tnlBICGSolver : public tnlMatrixSolver< T >
                const T* b,
                T* x, 
                const double& max_residue,
-               const long int max_iterations,
+               const int max_iterations,
                tnlPreconditioner< T >* precond = 0 )
    {
       if( ! SetSize( A. GetSize() ) ) return false;
@@ -54,7 +54,7 @@ template< typename T > class tnlBICGSolver : public tnlMatrixSolver< T >
       tnlMatrixSolver< T > :: iteration = 0;
       T alpha, beta, s1, s2;
       T b_norm( 0.0 );
-      long int i;
+      int i;
       for( i = 0; i < size; i ++ )
          b_norm += b[ i ] * b[ i ];
 
@@ -158,8 +158,8 @@ template< typename T > class tnlBICGSolver : public tnlMatrixSolver< T >
    {
       A. VectorProduct( x, tmp );
       T res = 0.0;
-      const long int size = A. GetSize();
-      long int i;
+      const int size = A. GetSize();
+      int i;
       for( i = 0; i < size; i ++ )
       {
          T v = tmp[ i ] - b[ i ];
@@ -168,7 +168,7 @@ template< typename T > class tnlBICGSolver : public tnlMatrixSolver< T >
       return sqrt( res ) / b_norm;
    };
 
-   bool AllocateSupportingArrays( long int size )
+   bool AllocateSupportingArrays( int size )
    {
       r = new T[ size ];
       r_ast = new T[ size ];
@@ -185,7 +185,7 @@ template< typename T > class tnlBICGSolver : public tnlMatrixSolver< T >
       return true;
    };
 
-   bool SetSize( long int _size )
+   bool SetSize( int _size )
    {
       if( size == _size ) return true;
       size = _size;
@@ -207,7 +207,7 @@ template< typename T > class tnlBICGSolver : public tnlMatrixSolver< T >
 
    tnlBaseMatrix< T >* A_T;
 
-   long int size;
+   int size;
 };
 
 #endif
